refactor: Repositories インターフェースを sessions/repositories 層に移動
sessions/manager.ts・engine-runner.ts が gateway/container.ts に依存する
sessions→gateway の逆方向依存を解消する。

- sessions/repositories/repositories.ts に Repositories 定義を移動
- gateway/container.ts は re-export のみ（後方互換維持）
- sessions/manager.ts・engine-runner.ts のインポートを sessions 層に変更
