+++ packages/jimmy/src/sessions/engine-runner.ts
@@ -1,6 +1,6 @@
 import fs from "node:fs";
 import { checkBudget } from "../gateway/budgets.js";
-import type { Repositories } from "../gateway/container.js";
+import type { Repositories } from "./repositories/index.js";
 import { scanOrg } from "../gateway/org.js";
 import { resolveOrgHierarchy } from "../gateway/org-hierarchy.js";
 import { cleanupMcpConfigFile, resolveMcpServers, writeMcpConfigFile } from "../mcp/resolver.js";
