+++ packages/jimmy/src/sessions/repositories/index.ts
@@ -8,6 +8,8 @@ export type { SessionMessage, IMessageRepository } from "./IMessageRepository.js
 export type { QueueItem, IQueueRepository } from "./IQueueRepository.js";
 export type { FileMeta, IFileRepository } from "./IFileRepository.js";
 
+export type { Repositories } from "./repositories.js";
+
 export { SqliteSessionRepository } from "./SqliteSessionRepository.js";
 export { SqliteMessageRepository } from "./SqliteMessageRepository.js";
 export { SqliteQueueRepository } from "./SqliteQueueRepository.js";
